requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.3.7